feat(protocol-designer): onboarding flow consolidate 1st 3 steps
