+++ protocol-designer/src/ProtocolRoutes.tsx
@@ -2,10 +2,10 @@ import { Route, Navigate, Routes, useNavigate } from 'react-router-dom'
 import { ErrorBoundary } from 'react-error-boundary'
 import { Box } from '@opentrons/components'
 import {
-  CreateNewProtocolWizard,
   Designer,
   Landing,
   Liquids,
+  Onboarding,
   ProtocolOverview,
   Settings,
 } from './pages'
@@ -40,7 +40,7 @@ const pdRoutes: RouteProps[] = [
     path: '/designer',
   },
   {
-    Component: CreateNewProtocolWizard,
+    Component: Onboarding,
     name: 'Create new protocol',
     navLinkTo: '/createNew',
     path: '/createNew',
