+++ protocol-designer/src/assets/localization/en/create_new_protocol.json
@@ -1,6 +1,9 @@
 {
   "add_custom_tips": "Add custom pipette tips",
   "add_fixtures": "Add your fixtures",
+  "add_your_pipettes": "Add your pipettes",
+  "are_you_using_thermocycler": "Are you using a Thermocycler in your protocol?",
+  "are_you_using_gripper": "Are you using a Waste chute in your protocol?",
   "add_gripper_for_absorbance_reader": "Add a gripper to enable adding Absorbance Plate Reader",
   "add_gripper": "Add a gripper",
   "add_modules": "Add your modules",
