+++ protocol-designer/src/components/organisms/SelectPipetteModal/__test__/SelectPipetteModal.test.tsx
@@ -2,24 +2,27 @@ import { describe, it, expect, vi, beforeEach } from 'vitest'
 import '@testing-library/jest-dom/vitest'
 import { FLEX_ROBOT_TYPE, OT2_ROBOT_TYPE } from '@opentrons/shared-data'
 import { fireEvent, screen } from '@testing-library/react'
-import { i18n } from '../../../assets/localization'
-import { renderWithProviders } from '../../../__testing-utils__'
-import { getLabwareDefsByURI } from '../../../labware-defs/selectors'
-import { getAllowAllTipracks } from '../../../feature-flags/selectors'
-import { IncompatibleTipsModal } from '../../../components/organisms'
-import { createCustomTiprackDef } from '../../../labware-defs/actions'
-import { SelectPipettes } from '../SelectPipettes'
-import { getTiprackOptions } from '../utils'
+import { getLabwareDefsByURI } from '../../../../labware-defs/selectors'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../assets/localization'
+import { getAllowAllTipracks } from '../../../../feature-flags/selectors'
+import { getTiprackOptions } from '../../../../pages/Onboarding/utils'
+import { createCustomTiprackDef } from '../../../../labware-defs/actions'
+import { IncompatibleTipsModal } from '../../IncompatibleTipsModal'
+import { SelectPipetteModal } from '..'
 
 import type { ComponentProps } from 'react'
 import type { NavigateFunction } from 'react-router-dom'
-import type { WizardFormState, WizardTileProps } from '../types'
+import type {
+  WizardFormState,
+  WizardTileProps,
+} from '../../../../pages/Onboarding/types'
 
-vi.mock('../../../labware-defs/selectors')
-vi.mock('../../../feature-flags/selectors')
-vi.mock('../../../components/organisms')
-vi.mock('../../../labware-defs/actions')
-vi.mock('../utils')
+vi.mock('../../../../labware-defs/actions')
+vi.mock('../../../../pages/Onboarding/utils')
+vi.mock('../../IncompatibleTipsModal')
+vi.mock('../../../../labware-defs/selectors')
+vi.mock('../../../../feature-flags/selectors')
 const mockLocation = vi.fn()
 window.HTMLElement.prototype.scrollIntoView = vi.fn()
 
@@ -31,8 +34,8 @@ vi.mock('react-router-dom', async importOriginal => {
   }
 })
 
-const render = (props: ComponentProps<typeof SelectPipettes>) => {
-  return renderWithProviders(<SelectPipettes {...props} />, {
+const render = (props: ComponentProps<typeof SelectPipetteModal>) => {
+  return renderWithProviders(<SelectPipetteModal {...props} />, {
     i18nInstance: i18n,
   })[0]
 }
@@ -45,7 +48,10 @@ const values = {
     organizationOrAuthor: '',
     robotType: FLEX_ROBOT_TYPE,
   },
-  pipettesByMount: { left: {}, right: {} },
+  pipettesByMount: {
+    left: { pipetteName: 'p1000_single_flex', tiprackDefURI: ['mockDefUri'] },
+    right: {},
+  },
   modules: null,
 } as WizardFormState
 
@@ -55,15 +61,22 @@ const mockWizardTileProps: Partial<WizardTileProps> = {
   watch: vi.fn((name: keyof typeof values) => values[name]) as any,
 }
 
-describe('SelectPipettes', () => {
-  let props: ComponentProps<typeof SelectPipettes>
+describe('SelectPipetteModal', () => {
+  let props: ComponentProps<typeof SelectPipetteModal>
 
   beforeEach(() => {
     props = {
       ...props,
-      goBack: vi.fn(),
       ...mockWizardTileProps,
-    } as WizardTileProps
+      handleBack: vi.fn(),
+      pipetteGen: 'flex',
+      pipetteVolume: 'p1000',
+      pipetteType: 'single',
+      setPipetteGen: vi.fn(),
+      setPipetteVolume: vi.fn(),
+      setPipetteType: vi.fn(),
+      mount: 'right',
+    }
     vi.mocked(IncompatibleTipsModal).mockReturnValue(
       <div>mock incompatible tips modal</div>
     )
@@ -77,11 +90,7 @@ describe('SelectPipettes', () => {
 
   it('renders the first page of select pipettes for a Flex', () => {
     render(props)
-    screen.getByText('Step 2')
     screen.getByText('Add a pipette')
-    screen.getByText(
-      'Pick your first pipette. If you need a second pipette, you can add it next.'
-    )
     screen.getByText('Pipette type')
     // select pip type
     fireEvent.click(screen.getByRole('label', { name: '1-Channel' }))
@@ -89,17 +98,16 @@ describe('SelectPipettes', () => {
     // select pip volume
     fireEvent.click(screen.getByRole('label', { name: '1000 µL' }))
     // select tip
-    screen.getByText('Add custom pipette tips')
     screen.getByText('200µL Flex tipracks')
     fireEvent.click(screen.getByText('1000µL Flex tipracks'))
 
-    screen.getByRole('button', { name: 'Confirm' })
+    screen.getByRole('button', { name: 'Save' })
 
     // go back
-    fireEvent.click(screen.getByRole('button', { name: 'Go back' }))
-    expect(props.goBack).toHaveBeenCalled()
+    fireEvent.click(screen.getByRole('button', { name: 'Cancel' }))
+    expect(props.handleBack).toHaveBeenCalled()
 
-    fireEvent.click(screen.getByRole('button', { name: 'Confirm' }))
+    fireEvent.click(screen.getByRole('button', { name: 'Save' }))
   })
 
   it('renders the first page of select pipettes for an ot-2', () => {
@@ -116,20 +124,29 @@ describe('SelectPipettes', () => {
         organizationOrAuthor: '',
         robotType: OT2_ROBOT_TYPE,
       },
-      pipettesByMount: { left: {}, right: {} },
+      pipettesByMount: {
+        left: {},
+        right: {
+          pipetteName: 'p20_single_gen2',
+          tiprackDefURI: ['mockDefUri'],
+        },
+      },
       modules: null,
     } as WizardFormState
 
     props = {
       ...props,
       watch: vi.fn((name: keyof typeof values) => values[name]) as any,
+      handleBack: vi.fn(),
+      pipetteGen: 'GEN2',
+      pipetteVolume: 'p20',
+      pipetteType: 'single',
+      setPipetteGen: vi.fn(),
+      setPipetteVolume: vi.fn(),
+      setPipetteType: vi.fn(),
+      mount: 'right',
     }
     render(props)
-    screen.getByText('Step 2')
-    screen.getByText('Add a pipette')
-    screen.getByText(
-      'Pick your first pipette. If you need a second pipette, you can add it next.'
-    )
     screen.getByText('Pipette type')
     // select pip type
     fireEvent.click(screen.getByRole('label', { name: '1-Channel' }))
