+++ protocol-designer/src/components/organisms/index.ts
@@ -22,6 +22,7 @@ export * from './LabwareUploadModal'
 export * from './Navigation'
 export * from './PipetteInfoItem'
 export * from './Portal'
+export * from './SelectPipetteModal'
 export * from './SelectWellsModal'
 export * from './SettingsIcon'
 export * from './SlotDetailsContainer'
