+++ protocol-designer/src/load-file/types.ts
@@ -21,7 +21,7 @@ export interface NewProtocolFields {
   name: string | null | undefined
   description: string | null | undefined
   organizationOrAuthor: string | null | undefined
-  robotType: RobotType
+  robotType?: RobotType
 }
 export interface LoadFileAction {
   type: 'LOAD_FILE'
