+++ protocol-designer/src/pages/Onboarding/ModuleDiagram.tsx
