+++ protocol-designer/src/pages/Onboarding/__tests__/SelectFixtures.test.tsx
